added a sophisticated theme gen to make everything look coherent between our different tools

+++ fish/functions/clean_nvim_swap.fish
@@ -3,8 +3,10 @@ function clean_nvim_swap --description 'Clean Neovim swap files'
 
   if test -d $swap_dir
     set -l count (count $swap_dir/*)
-    rm -f $swap_dir/*
-    echo "Cleaned $count swap files from Neovim swap directory"
+    if test $count -gt 0
+      rm -f $swap_dir/*
+      echo "Cleaned $count swap files from Neovim swap directory"
+    end
   else
     echo "Neovim swap directory not found at $swap_dir"
   end
